+++ tests/data/included/shire/tests/functional/TestGuy.php
@@ -302,7 +302,7 @@ class TestGuy extends \Codeception\Actor
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
-     * Checks if file does not exists in path
+     * Checks if file does not exist in path
      *
      * @param $filename
      * @param string $path
@@ -315,7 +315,7 @@ class TestGuy extends \Codeception\Actor
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
-     * Checks if file does not exists in path
+     * Checks if file does not exist in path
      *
      * @param $filename
      * @param string $path